add init to Date:  millisecondsSince1970
This convenience initializer take milliseconds as Long instead of a Double value of seconds. This is similar to the Date class in Javascript, and is useful for many JSON APIs that use number of milliseconds since epoch to represent dates.

Further I think the word TimeInterval should be renamed to Seconds in the API of the Date object to be less vague.